Link itemsPerPage and number of lines in table.
Fix a case whennumber of distinct values is greater than itemsPerPage, they are not all displayed.
With a value of zero, the limit() function of mongoDB should apply no limit (what we want), but
other tests are false, so itemsPerPage is not taking into account as a /compute parameter, when set
to zero.

Temporary solution: ask for 100 itemsPerPage in /compute.json?o=distinct